service/pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = ""
authors = ["Matt McCredie <mccredie+git@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-jose = "^3.1.0"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]
boto3 = "^1.12.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
